+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - MAJOR This refactoring changes the way the h-adaptivity is decided for the .prm by changing the `kelly` `type` to `adaptive` and adding a new parameter in the mesh adaptation subsection : `error estimator`. This new parameter follow the same logic as the variable parameter and therefore enable the use of different types of error estimators for different physics. This refactor was necessary for the Time-harmonic Maxwell solver that has a DPG built-in error indicator. This new error indicator is also introduced in the PR along with h-adaptivity for the time-harmonic auxiliary physics. Note that I also needed to change the type of element from FE_NedelecSZ to FE_Nedelec, because there is a bug with the SZ version when using multiple core in deal.ii library currently.[#1922](https://github.com/chaos-polymtl/lethe/pull/1922)
 
+- MAJOR The CFD-DEM solvers (`CFDDEMSolver` and `CFDDEMMatrixFree`) have been refactored to centralize DEM sub-iteration management within the `SubSimulationControlDEM` class. The `coupling_frequency` and `dem_time_step`  member variables have been removed and replaced by a shared `SubSimulationControlDEM` object from which the state of the DEM steps is queried. This also fixes a bug in which particle location within the cells would not be forced at the last DEM time-step if the DEM time step was based on a fraction of the Rayleigh time step. [#1929](https://github.com/chaos-polymtl/lethe/pull/1929)
+
 ## [Master] - 2026/03/03
 
 ### Added